chore: move autoplay functionality to hook
