+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-07-04T15:00:39.114Z\n"
-"PO-Revision-Date: 2025-07-04T15:00:39.114Z\n"
+"POT-Creation-Date: 2025-08-05T12:54:44.743Z\n"
+"PO-Revision-Date: 2025-08-05T12:54:44.743Z\n"
 
 msgid "Untitled dashboard"
 msgstr "Untitled dashboard"
@@ -712,24 +712,6 @@ msgstr ""
 msgid "Yes, remove filters"
 msgstr "Yes, remove filters"
 
-msgid "5 seconds per slide"
-msgstr "5 seconds per slide"
-
-msgid "10 seconds per slide"
-msgstr "10 seconds per slide"
-
-msgid "20 seconds per slide"
-msgstr "20 seconds per slide"
-
-msgid "30 seconds per slide"
-msgstr "30 seconds per slide"
-
-msgid "1 minute per slide"
-msgstr "1 minute per slide"
-
-msgid "2 minutes per slide"
-msgstr "2 minutes per slide"
-
 msgid "Exit slideshow"
 msgstr "Exit slideshow"
 
@@ -789,3 +771,21 @@ msgstr "This dashboard cannot be loaded while offline."
 
 msgid "Go to start page"
 msgstr "Go to start page"
+
+msgid "5 seconds per slide"
+msgstr "5 seconds per slide"
+
+msgid "10 seconds per slide"
+msgstr "10 seconds per slide"
+
+msgid "20 seconds per slide"
+msgstr "20 seconds per slide"
+
+msgid "30 seconds per slide"
+msgstr "30 seconds per slide"
+
+msgid "1 minute per slide"
+msgstr "1 minute per slide"
+
+msgid "2 minutes per slide"
+msgstr "2 minutes per slide"
